+++ src/commands/mod.rs
@@ -1,0 +1,11 @@
+pub mod add;
+pub mod commit;
+pub mod init;
+pub mod log;
+pub mod push;
+
+pub use add::git_add;
+pub use commit::git_commit;
+pub use init::git_init;
+pub use log::git_log;
+pub use push::git_push;
