+++ src/main.rs
@@ -1,84 +1,16 @@
-use git2::{IndexAddOption, Repository};
+mod commands;
+
+use git2::Repository;
 use std::io::{self, BufRead, Write};
 use std::path::Path;
-
-pub fn git_init() -> Result<(), git2::Error> {
-    let _repo = Repository::init(".")?;
-    println!("repo init success.");
-    Ok(())
-}
-
-pub fn git_add(path_str: &str) -> Result<(), git2::Error> {
-    let repo = Repository::open(".")?;
-    let mut idx = repo.index()?;
-
-    if path_str == "." {
-        idx.add_all(["."].iter(), IndexAddOption::DEFAULT, None)?;
-        println!("all added.");
-    } else {
-        let path = Path::new(path_str);
-        idx.add_path(&path)?;
-        println!("{} added.", path_str);
-    }
-    idx.write()?;
-    Ok(())
-}
-
-pub fn git_log() -> Result<Vec<String>, git2::Error> {
-    let repo = Repository::open(".")?;
-    let mut revwalk = repo.revwalk()?;
-    revwalk.push_head()?;
-
-    let mut res = Vec::new();
-    for commit_id in revwalk {
-        let commit = repo.find_commit(commit_id?)?;
-        if let Some(msg) = commit.message() {
-            res.push(msg.to_string());
-        }
-    }
-    Ok(res)
-}
-
-fn git_commit(message: &str) -> Result<(), git2::Error> {
-    let repo = Repository::open(".")?;
-    let mut idx = repo.index()?;
-
-    let tree_id = idx.write_tree()?;
-    let tree = repo.find_tree(tree_id)?;
-    let sig = repo.signature()?;
-
-    let parent_commits = match repo.head() {
-        Ok(head_ref) => {
-            let head = head_ref
-                .target()
-                .ok_or_else(|| git2::Error::from_str("HEAD refers to non-HEAD"))?;
-            vec![repo.find_commit(head)?]
-        }
-        Err(_) => Vec::new(),
-    };
-
-    let parents: Vec<&git2::Commit> = parent_commits.iter().collect();
-
-    let commit_oid = repo.commit(Some("HEAD"), &sig, &sig, message, &tree, &parents)?;
-    println!("commit created: {}", commit_oid);
-    Ok(())
-}
-
-pub fn git_push(remote_name: &str, refspec: &str) -> Result<(), git2::Error> {
-    let repo = Repository::open(".")?;
-    let mut remote = repo.find_remote(remote_name)?;
-
-    remote.push(&[refspec], None)?;
-    println!("push complete to remote: {}", remote_name);
-    Ok(())
-}
 
 pub fn main() -> Result<(), git2::Error> {
     let stdin = io::stdin();
     let mut stdout = io::stdout();
 
     loop {
-        print!("명령어 입력 (init, add <path>, commit <msg>, q): ");
+        // todo --help 구현
+        print!("명령어 입력 (init, add <path>, commit <msg>, log, push <remote> <refspec>, q): ");
         stdout.flush().unwrap();
         let mut input = String::new();
         stdin.lock().read_line(&mut input).unwrap();
@@ -91,7 +23,7 @@ pub fn main() -> Result<(), git2::Error> {
 
         match tokens[0] {
             "init" => {
-                if let Err(e) = git_init() {
+                if let Err(e) = commands::git_init() {
                     println!("init error: {}", e);
                 }
             }
@@ -99,7 +31,7 @@ pub fn main() -> Result<(), git2::Error> {
                 if tokens.len() < 2 {
                     println!("input file path");
                 } else {
-                    if let Err(e) = git_add(tokens[1]) {
+                    if let Err(e) = commands::git_add(tokens[1]) {
                         println!("add error: {}", e);
                     }
                 }
@@ -109,7 +41,7 @@ pub fn main() -> Result<(), git2::Error> {
                     println!("input commit message");
                 } else {
                     let commit_msg = tokens[1..].join(" ");
-                    if let Err(e) = git_commit(&commit_msg) {
+                    if let Err(e) = commands::git_commit(&commit_msg) {
                         println!("commit error: {}", e);
                     }
                 }
@@ -120,12 +52,12 @@ pub fn main() -> Result<(), git2::Error> {
                 } else {
                     let remote = tokens[1];
                     let refspec = tokens[2];
-                    if let Err(e) = git_push(remote, refspec) {
+                    if let Err(e) = commands::git_push(remote, refspec) {
                         println!("push error: {}", e);
                     }
                 }
             }
-            "log" => match git_log() {
+            "log" => match commands::git_log() {
                 Ok(logs) => {
                     println!("커밋 로그:");
                     for msg in logs {
@@ -156,12 +88,12 @@ mod tests {
     static SHARED_REPO: Lazy<TempDir> = Lazy::new(|| {
         let tmp_dir = TempDir::new().expect("failed to create temporary directory");
         env::set_current_dir(tmp_dir.path()).expect("failed to set temporary directory");
-        git_init().expect("failed to git init");
+        commands::git_init().expect("failed to git init");
         tmp_dir
     });
 
     fn get_repo() -> Repository {
-        let _ = &*SHARED_REPO;
+        env::set_current_dir(SHARED_REPO.path()).expect("failed to set temporary directory");
         Repository::open(".").expect("failed to open repo")
     }
 
@@ -182,7 +114,7 @@ mod tests {
         let file_path = Path::new(file_name.as_str());
         File::create(file_path).expect("failed to create temp file");
 
-        git_add(file_name.as_str()).expect("failed to add file");
+        commands::git_add(file_name.as_str()).expect("failed to add file");
 
         let index = repo.index().expect("failed to get the index");
         let entries: Vec<_> = index
@@ -201,7 +133,7 @@ mod tests {
         let file_path = Path::new(file_name.as_str());
         File::create(file_path).expect("failed to create temp file");
 
-        git_add(".").expect("failed to add file");
+        commands::git_add(".").expect("failed to add file");
 
         let index = repo.index().unwrap();
         let entries: Vec<_> = index
@@ -219,10 +151,10 @@ mod tests {
         let file_name = generate_random_file_name(".txt");
         let file_path = Path::new(file_name.as_str());
         File::create(file_path).expect("failed to create temp file");
-        git_add(file_name.as_str()).expect("failed to add file");
+        commands::git_add(file_name.as_str()).expect("failed to add file");
 
         let commit_msg = "test commit msg";
-        git_commit(commit_msg).expect("failed to commit message");
+        commands::git_commit(commit_msg).expect("failed to commit message");
 
         let head_commit = {
             let head = repo.head().expect("failed to get HEAD");
@@ -245,13 +177,13 @@ mod tests {
         let file_name = generate_random_file_name(".txt");
         let file_path = Path::new(file_name.as_str());
         File::create(file_path).expect("failed to create temp file");
-        git_add(file_name.as_str()).expect("failed to add file");
+        commands::git_add(file_name.as_str()).expect("failed to add file");
 
         let commit_msg = "log test commit";
 
-        git_commit(commit_msg).expect("failed to commit");
+        commands::git_commit(commit_msg).expect("failed to commit");
 
-        let logs = git_log().expect("failed to get log");
+        let logs = commands::git_log().expect("failed to get log");
         assert_eq!(logs.first().unwrap().trim(), commit_msg, "커밋 로그가 다름");
     }
 
@@ -298,7 +230,7 @@ mod tests {
         }
 
         // 로컬에서 원격의 refs/heads/main로 push 수행
-        git_push("origin", "refs/heads/main").expect("failed to push origin");
+        commands::git_push("origin", "refs/heads/main").expect("failed to push origin");
 
         // 원격 repo에서 HEAD commit 검증
         let remote_head = remote_repo
